refactored deployer
extracted "depinfo" delpoyment-info tool
