+++ .gitignore
@@ -47,3 +47,4 @@ hardhat.log
 /packages/contracts/docs/
 /packages/deployer/contracts-link
 /packages/deployer/all-networks.json
+/packages/deployer/deployments/
